requirements.txt
google-adk
python-dotenv
openai
requests
litellm 